% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corPlot_matchall}
\alias{corPlot_matchall}
\title{Plot heatmap}
\usage{
corPlot_matchall(
  beta,
  beta_true,
  theta,
  theta_true,
  title = NULL,
  annotation = TRUE
)
}
\description{
This function plots both beta theta correlation heatmaps together.
}
