% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PerplexityPlot}
\alias{PerplexityPlot}
\title{Perplexity plot}
\usage{
PerplexityPlot(models, corpus = NULL, perc.rare.thresh = 0.05)
}
\description{
This function plots perplexity and number of rare cell type.
}
