% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vizAllTopics_adjust}
\alias{vizAllTopics_adjust}
\title{Plot Piechart}
\usage{
vizAllTopics_adjust(
  theta,
  pos,
  topicOrder = seq(ncol(theta)),
  topicCols = rainbow(ncol(theta)),
  groups = NA,
  group_cols = NA,
  r = max(0.4, max(pos)/nrow(pos) * 4),
  lwd = 0.01,
  showLegend = TRUE,
  plotTitle = NA,
  overlay = NA
)
}
\description{
This function plots theta piechart.
}
