% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vizGenesCounts}
\alias{vizGenesCounts}
\title{Gene visualization}
\usage{
vizGenesCounts(
  counts,
  spatial_location,
  genes,
  groups = NA,
  group_cols = NA,
  size = 7,
  stroke = 0.5,
  alpha = 1,
  plotTitle = NA,
  showLegend = TRUE
)
}
\description{
This function plots gene expression.
}
