% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLDA.R
\name{WLDA}
\alias{WLDA}
\title{weight LDA}
\usage{
WLDA(
  corpus,
  k,
  type = "bdc",
  seed = 0,
  iterations = 1500,
  priors = NULL,
  beta_init = NULL,
  ldamodel = NULL,
  verbose = FALSE
)
}
\description{
This function do the deconvolution with different weight schemes.
}
